feat: Create the forms for AddSection
Create the form with uncontrolled inputs because having controlled
inputs doesn't allow us to clear it in any way. Because with controlled
input you can't use the reset method of a form.

Used the uuid package to assign a unique unchangeable id because we
can't use array ids since they will be changing with every filter and
delete of an entry(On the other hand while creating the fields we use
array indexes because they will be constant all the time).