fix: Add controlled inputs directly in component
FINALLY DID IT.
Previously was struggling to implement it because of a warning that
appeared because there was no initial value, i.e. rendering the input as
uncontrolled, then upon change changes it to controlled with state. The
problem was probably that the initial state was being complied
incorrectly(`[].fill("", 0, 2)` - doesn't work correctly because there
are no elements in the array).