feat: Add functionality for editing sections
Added an edit property when creating entry for information about
editing.

Passed in fields in the ItemsList components because it is needed for
the new EditForm component.

Based on the new edit property, a list item is shown or the EditForm.

The new EditForm component has the ability to cancel an edit but more
importantly edit an entry.